backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.5
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tqdm==4.64.1
twilio==7.16.3
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
